+++ packages/iconsax/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 0.2.0 (2022-05-07)
+# [0.7.0](https://github.com/MM25Zamanian/mmzmk/compare/v0.6.0...v0.7.0) (2022-05-09)
 
 **Note:** Version bump only for package @mmzmk/iconsax
+
+
+
+
+
+# [0.6.0](https://github.com/MM25Zamanian/mmzmk/compare/v0.5.1...v0.6.0) (2022-05-09)
+
+**Note:** Version bump only for package @mmzmk/iconsax
+
+
+
+
+
+# [0.5.0](https://github.com/MM25Zamanian/mmzmk/compare/v0.4.1...v0.5.0) (2022-05-09)
+
+
+### Bug Fixes
+
+* **iconsax:** fetch url ([#7](https://github.com/MM25Zamanian/mmzmk/issues/7)) ([e278505](https://github.com/MM25Zamanian/mmzmk/commit/e2785057be339559573e94e21e03c8e6b77215ce))
+
+
+
+
+
+# [0.4.0](https://github.com/MM25Zamanian/mmzmk/compare/v0.3.1...v0.4.0) (2022-05-09)
+
+
+### Features
+
+* **iconsax:** add Crypto to icon name list type ([4f994a5](https://github.com/MM25Zamanian/mmzmk/commit/4f994a526659ab7da03518a8d665d8a9598f3a7a))
+
+
+
+
+
+# 0.3.0 (2022-05-08)
+
+
+### Bug Fixes
+
+* **iconsax:** fetch url ([08ad5a2](https://github.com/MM25Zamanian/mmzmk/commit/08ad5a23d8b79bc694f7a798498feacf8f05c5be))
+
+
+### Features
+
+* **iconsax:** add jsdocs example ([fae6254](https://github.com/MM25Zamanian/mmzmk/commit/fae6254fa81d4f423575096a9e5084235a000226))